+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 * Updated dependencies
 
 ### Fixed
-* Querying trips with limit directly from `Query` instance (#18)
+* Querying trips and messages with limit directly from `Query` instance (#18)
 
 ### Misc
 * Tested with JDK 19